Correções e ajuste de arquitetura
