+++ SME.Pedagogico.Gestao.Data/Business/PollMatematica.cs
@@ -33,21 +33,11 @@ namespace SME.Pedagogico.Gestao.Data.Business
                 {
                     case "1° Semestre":
                         listStudentsPoll = db.MathPoolCAs.Where(x => x.TurmaEolCode == turmaEol &&
-                                x.Semestre.Equals(1) &&
-                                (!string.IsNullOrEmpty(x.Ordem1Ideia) ||
-                                !string.IsNullOrEmpty(x.Ordem2Ideia) ||
-                                !string.IsNullOrEmpty(x.Ordem3Ideia) ||
-                                !string.IsNullOrEmpty(x.Ordem4Ideia)
-                                )).ToList();
+                                x.Semestre.Equals(1)).ToList();
                         break;
                     case "2° Semestre":
                         listStudentsPoll = db.MathPoolCAs.Where(x => x.TurmaEolCode == turmaEol &&
-                                x.Semestre.Equals(2) &&
-                                (!string.IsNullOrEmpty(x.Ordem1Ideia) ||
-                                !string.IsNullOrEmpty(x.Ordem2Ideia) ||
-                                !string.IsNullOrEmpty(x.Ordem3Ideia) ||
-                                !string.IsNullOrEmpty(x.Ordem4Ideia)
-                                )).ToList();
+                                x.Semestre.Equals(2)).ToList();
                         break;
                 }
                 return listStudentsPoll;
@@ -64,25 +54,11 @@ namespace SME.Pedagogico.Gestao.Data.Business
                 {
                     case "1° Semestre":
                         listStudentsPoll = db.MathPoolCMs.Where(x => x.TurmaEolCode == turmaEol &&
-                        x.Semestre.Equals(1) &&
-                            (!string.IsNullOrEmpty(x.Ordem3Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem4Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem5Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem6Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem7Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem8Ideia)
-                            )).ToList();
+                        x.Semestre.Equals(1)).ToList();
                         break;
                     case "2° Semestre":
                         listStudentsPoll = db.MathPoolCMs.Where(x => x.TurmaEolCode == turmaEol &&
-                        x.Semestre.Equals(2) &&
-                            (!string.IsNullOrEmpty(x.Ordem3Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem4Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem5Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem6Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem7Ideia)
-                            || !string.IsNullOrEmpty(x.Ordem8Ideia)
-                            )).ToList();
+                       x.Semestre.Equals(2)).ToList();
                         break;
 
                 }
@@ -100,28 +76,12 @@ namespace SME.Pedagogico.Gestao.Data.Business
                 {
                     case "1° Semestre":
                         listStudentsPoll = db.MathPoolNumbers.Where(x => x.TurmaEolCode == turmaEol
-                                          && x.Semestre.Equals(1) &&
-                                             (!string.IsNullOrEmpty(x.Familiares)
-                                             || !string.IsNullOrEmpty(x.Opacos)
-                                             || !string.IsNullOrEmpty(x.Transparentes)
-                                             || !string.IsNullOrEmpty(x.TerminamZero)
-                                             || !string.IsNullOrEmpty(x.Algarismos)
-                                             || !string.IsNullOrEmpty(x.Processo)
-                                             || !string.IsNullOrEmpty(x.ZeroIntercalados)
-                                             )).ToList();
+                                          && x.Semestre.Equals(1)).ToList();
 
                         break;
                     case "2° Semestre":
                         listStudentsPoll = db.MathPoolNumbers.Where(x => x.TurmaEolCode == turmaEol &&
-                                           x.Semestre.Equals(2) &&
-                                             (!string.IsNullOrEmpty(x.Familiares)
-                                             || !string.IsNullOrEmpty(x.Opacos)
-                                             || !string.IsNullOrEmpty(x.Transparentes)
-                                             || !string.IsNullOrEmpty(x.TerminamZero)
-                                             || !string.IsNullOrEmpty(x.Algarismos)
-                                             || !string.IsNullOrEmpty(x.Processo)
-                                             || !string.IsNullOrEmpty(x.ZeroIntercalados)
-                                             )).ToList();
+                                           x.Semestre.Equals(2)).ToList();
 
                         break;
                     default:
